backend/ml/requirements.txt

flask-cors
numpy
google-cloud-logging  
pandas
scikit-learn
pickle-mixin
flask

